remove SafeList route because the list moved to sidebar
