+++ src/components/Sidebar/index.jsx
@@ -13,7 +13,7 @@ import Spacer from '~/components/Spacer'
 import Hairline from '~/components/layout/Hairline'
 import Row from '~/components/layout/Row'
 import { type Safe } from '~/routes/safe/store/models/safe'
-import { safesListSelector, defaultSafeSelector } from '~/routes/safeList/store/selectors'
+import { safesListSelector, defaultSafeSelector } from '~/routes/safe/store/selectors'
 import setDefaultSafe from '~/routes/safe/store/actions/setDefaultSafe'
 import useSidebarStyles from './style'
 import SafeList from './SafeList'
