+++ src/routes/safe/store/actions/addSafe.js
@@ -3,7 +3,7 @@ import { List } from 'immutable'
 import { createAction } from 'redux-actions'
 import type { Dispatch as ReduxDispatch, GetState } from 'redux'
 import { type GlobalState } from '~/store'
-import { safesListSelector } from '~/routes/safeList/store/selectors'
+import { safesListSelector } from '~/routes/safe/store/selectors'
 import { type Safe } from '~/routes/safe/store/models/safe'
 import { makeOwner } from '~/routes/safe/store/models/owner'
 import setDefaultSafe from '~/routes/safe/store/actions/setDefaultSafe'
