+++ src/routes/safe/store/middleware/safeStorage.js
@@ -12,7 +12,7 @@ import { type GlobalState } from '~/store/'
 import {
   saveSafes, setOwners, removeOwners, saveDefaultSafe,
 } from '~/logic/safe/utils'
-import { safesMapSelector } from '~/routes/safeList/store/selectors'
+import { safesMapSelector } from '~/routes/safe/store/selectors'
 import { getActiveTokensAddressesForAllSafes } from '~/routes/safe/store/selectors'
 import { tokensSelector } from '~/logic/tokens/store/selectors'
 import type { Token } from '~/logic/tokens/store/model/token'
