--- src/routes/safeList/container/index.jsx
@@ -1,21 +1,0 @@
-// @flow
-import { List } from 'immutable'
-import * as React from 'react'
-import { connect } from 'react-redux'
-import Page from '~/components/layout/Page'
-import { type Safe } from '~/routes/safe/store/models/safe'
-import Layout from '../components/Layout'
-import selector from './selector'
-
-type Props = {
-  safes: List<Safe>,
-  provider: string,
-}
-
-const SafeList = ({ safes, provider }: Props) => (
-  <Page>
-    <Layout safes={safes} provider={provider} />
-  </Page>
-)
-
-export default connect<*, *, *, *>(selector)(SafeList)
